feat: исправлен функционал бека: у каждого юзера свои заметки

+++ server/controllers/notes-controller.js
@@ -1,30 +1,56 @@
 import Note from "../models/note-model.js";
 
-async function addNote(note) {
-  const newNote = await Note.create(note);
+async function addNote(note, userId) {
+  console.log("addNote", note, userId);
+  const newNote = await Note.create({ ...note, userId });
+
   return newNote;
 }
 
-async function editNote(id, newData) {
-  console.log("controller", id, newData); //выходят верные данные
-  try {
-    const updatedNote = await Note.findByIdAndUpdate(id, newData, {
-      returnDocument: "after",
-    });
-    console.log(updatedNote); //выходит null
-    return updatedNote;
-  } catch (error) {
+async function editNote(id, newData, userId) {
+  const updatedNote = await Note.findOneAndUpdate(
+    { _id: id, userId }, // Проверяем, что заметка принадлежит пользователю
+    newData,
+    { new: true }
+  );
+  if (!updatedNote) {
     throw new Error("Не найдено");
   }
+  return updatedNote;
 }
 
-async function deleteNote(id) {
-  return await Note.deleteOne({ _id: id });
+async function deleteNote(id, userId) {
+  return await Note.deleteOne({ _id: id, userId }); // Проверяем, что заметка принадлежит пользователю
 }
 
-async function getNotes(search = "", limit, page) {
+// async function getNotes(search = "", limit, page) {
+//   const [notes, count] = await Promise.all([
+//     Note.find({
+//       $or: [
+//         { title: { $regex: search, $options: "i" } },
+//         { text: { $regex: search, $options: "i" } },
+//       ],
+//     })
+//       .limit(limit)
+//       .skip((page - 1) * limit)
+//       .sort({ createdAt: -1 }),
+//     Note.countDocuments({
+//       $or: [
+//         { title: { $regex: search, $options: "i" } },
+//         { text: { $regex: search, $options: "i" } },
+//       ],
+//     }),
+//   ]);
+//   return {
+//     notes,
+//     lastPage: Math.ceil(count / limit),
+//   };
+// }
+
+async function getNotes(userId, search = "", limit, page) {
   const [notes, count] = await Promise.all([
     Note.find({
+      userId, // Фильтруем заметки по userId
       $or: [
         { title: { $regex: search, $options: "i" } },
         { text: { $regex: search, $options: "i" } },
@@ -34,6 +60,7 @@ async function getNotes(search = "", limit, page) {
       .skip((page - 1) * limit)
       .sort({ createdAt: -1 }),
     Note.countDocuments({
+      userId, // Фильтруем по userId
       $or: [
         { title: { $regex: search, $options: "i" } },
         { text: { $regex: search, $options: "i" } },
